requirements.txt
tensorflow-gpu==2.12.0
scikit-learn==1.2.2
matplotlib==3.7.1
numpy==1.24.2
pillow==9.5.0
